moving fetch and colors file
