+++ components/Navbar.vue
@@ -127,8 +127,8 @@ import { Component, Vue } from 'nuxt-property-decorator'
 import LocaleChanger from '@/components/shared/SwitchLocale.vue'
 import HistoryBrowser from '@/components/shared/history/HistoryBrowser.vue'
 import NavbarProfileDropdown from '@/components/rmrk/Profile/NavbarProfileDropdown.vue'
-import { getCurrentColor } from '@/colors'
 import i18n from '@/i18n'
+import { getCurrentColor } from '@/utils/colors'
 
 @Component({
   components: {
