+++ plugins/globalVariables.ts
@@ -7,7 +7,7 @@ import { enableExtension } from '@/extension'
 import { web3FromAddress } from '@polkadot/extension-dapp'
 import 'setimmediate'
 import mingo from 'mingo'
-import api from '@/fetch'
+import api from '@/utils/fetch'
 import { baseIpfsPrice, cost,  getFileSize, supportTx } from '@/utils/support'
 import axios from 'axios'
 import { set, get, getMany } from 'idb-keyval'
